+++ include/TPP/Dialect/LinalgX/CMakeLists.txt
@@ -1,5 +1,3 @@
 add_mlir_dialect(LinalgXOps linalgx)
 add_mlir_doc(LinalgXDialect LinalgXDialect TPP/ -gen-dialect-doc)
 add_mlir_doc(LinalgXOps LinalgXOps TPP/ -gen-op-doc)
-
-add_subdirectory(TransformOps)
