+++ lib/TPP/Dialect/CMakeLists.txt
@@ -3,3 +3,4 @@ add_subdirectory(Tpp)
 add_subdirectory(Xsmm)
 add_subdirectory(Check)
 add_subdirectory(VNNI)
+add_subdirectory(Transform)
